Struct mirrord_operator::client::OperatorApi

pub struct OperatorApi {
    client: Client,
    target_api: Api<TargetCrd>,
    copy_target_api: Api<CopyTargetCrd>,
    target_namespace: Option<String>,
    target_config: TargetConfig,
    on_concurrent_steal: ConcurrentSteal,
}

Fields
client: Clienttarget_api: Api<TargetCrd>copy_target_api: Api<CopyTargetCrd>target_namespace: Option<String>target_config: TargetConfigon_concurrent_steal: ConcurrentSteal
Implementations

impl OperatorApi

const COPIED_POD_IDLE_TTL: u32 = 30u32

We allow copied pods to live only for 30 seconds before the internal proxy connects.

fn check_config( config: &LayerConfig, operator: &MirrordOperatorCrd ) -> Result<(), OperatorApiError>

Checks used config against operator specification.

pub async fn create_session<P>( config: &LayerConfig, progress: &P, analytics: &mut AnalyticsReporter ) -> Result<Option<OperatorSessionConnection>, OperatorApiError>where P: Progress + Send + Sync,

Creates new OperatorSessionConnection based on the given LayerConfig. Returns None if the operator is not found.

pub async fn connect( config: &LayerConfig, session_information: OperatorSessionInformation, analytics: Option<&mut AnalyticsReporter> ) -> Result<OperatorSessionConnection, OperatorApiError>

Connects to exisiting operator session based on the given LayerConfig and OperatorSessionInformation.

async fn new(config: &LayerConfig) -> Result<Self, OperatorApiError>

async fn fetch_operator( &self ) -> Result<Option<MirrordOperatorCrd>, OperatorApiError>

async fn fetch_target(&self) -> Result<TargetCrd, OperatorApiError>

fn namespace(&self) -> &str

Returns a namespace of the target.

fn connect_url(&self, session: &OperatorSessionInformation) -> String

Returns a connection url for the given OperatorSessionInformation. This can be used to create a websocket connection with the operator.

async fn check_no_port_locks( &self, target: &TargetCrd ) -> Result<(), OperatorApiError>

Checks that there are no active port locks on the given target.

async fn connect_target( &self, session_info: OperatorSessionInformation ) -> Result<OperatorSessionConnection, OperatorApiError>

Create websocket connection to operator.

async fn copy_target( &self, session_metadata: &OperatorSessionMetadata, target: TargetCrd, scale_down: bool ) -> Result<CopyTargetCrd, OperatorApiError>

Creates a new CopyTargetCrd resource using the operator. This should create a new dummy pod out of the given target.